requirements-streaming.txt
numpy
opencv-python
pyserial
argparse
coloredlogs
bleak